+++ gems/aws-sdk-s3/lib/aws-sdk-s3/presigner.rb
@@ -231,17 +231,23 @@ module Aws
           end
           http_req.endpoint.query = query.join('&') unless query.empty?
 
+          signing_algorithm = :sigv4
+
           # If it's an ARN, get the resolved region and service
           if (arn = context.metadata[:s3_arn])
             region = arn[:resolved_region]
             service = arn[:arn].service
+            region = arn[:arn].is_a?(MultiRegionAccessPointARN) ? '*': arn[:resolved_region]
+            signing_algorithm = arn[:arn].is_a?(MultiRegionAccessPointARN) ? :sigv4a : :sigv4
           end
 
           signer = Aws::Sigv4::Signer.new(
             service: service || 's3',
             region: region || context.config.region,
+            signing_algorithm: signing_algorithm,
             credentials_provider: context.config.credentials,
             unsigned_headers: unsigned_headers,
+            apply_checksum_header: false,
             uri_escape_path: false
           )